create and integrate utils file
